+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/decorators/Top.java
@@ -1,0 +1,6 @@
+package com.github.thesench.solr.dsl.stream.expr.decorators;
+
+public class Top {
+    private Top() {}
+    
+}
